Support for choosing serial port everywhere

+++ resources/tools/fs/flashFs.sh
@@ -8,6 +8,13 @@ offset=$(<in/offset.txt tr -d '\n')
 # https://electronics.stackexchange.com/questions/588293/esp32-littlefs-change-block-size
 # Fuck
 
+source ../globalFunctions.sh
+serial_port=$(extract_serial_port "../other/in/esptool")
+if [[ $? -ne 0 || -z "$serial_port" ]]; then
+    echo "Failed to detect a valid serial port" >&2
+    exit 1
+fi
+
 rm -rf out/fs.bin
 mv ./littlefs/other/yatchy-lp-program*.bin ./littlefs/other/yatchy-lp-program.bin 2>/dev/null || true
 ./in/mklittlefs --all-files -c littlefs -s $size out/fs.bin
@@ -30,4 +37,4 @@ fi
 # sync
 # sudo losetup -d /dev/loop84
 
-../other/in/esptool write_flash $offset out/fs.bin
+../other/in/esptool --port $serial_port write_flash $offset out/fs.bin
